refactor(FormConfiguration): move FormConfiguration to its own file and update imports
