+++ server/thryve/src/main/kotlin/com/hatchgrid/thryve/form/domain/Form.kt
@@ -1,7 +1,6 @@
 package com.hatchgrid.thryve.form.domain
 
 import com.hatchgrid.common.domain.BaseEntity
-import com.hatchgrid.thryve.form.domain.dto.FormConfiguration
 import com.hatchgrid.thryve.form.domain.event.FormCreatedEvent
 import com.hatchgrid.thryve.form.domain.event.FormUpdatedEvent
 import com.hatchgrid.thryve.workspace.domain.WorkspaceId
