+++ server/thryve/src/test/kotlin/com/hatchgrid/thryve/form/domain/FormTest.kt
@@ -1,7 +1,6 @@
 package com.hatchgrid.thryve.form.domain
 
 import com.hatchgrid.UnitTest
-import com.hatchgrid.thryve.form.domain.dto.FormConfiguration
 import com.hatchgrid.thryve.form.domain.event.FormCreatedEvent
 import com.hatchgrid.thryve.form.domain.event.FormUpdatedEvent
 import java.time.LocalDateTime
